Thang

+++ lib/app/json_to_dart/order/customer_order/list_cus_order.dart
@@ -313,7 +313,7 @@ class WorkerInOrders {
   int? id;
   int? houseWorkerId;
   int? orderId;
-  int? rating;
+  double? rating;
   HouseWorker? houseWorker;
   String? order;
